Conflicts:
	src/org/usfirst/frc/team5026/robot/PantherJoystick.java
	src/org/usfirst/frc/team5026/robot/RobotMap.java

+++ src/org/usfirst/frc/team5026/robot/Constants.java
@@ -1,7 +1,7 @@
 package org.usfirst.frc.team5026.robot;
 
 public class Constants {
-	public static final double X_SCALING = 0.6;
-	public static final double Y_SCALING = 0.75; 
+	public static final float X_SCALING = 0.6f;
+	public static final float Y_SCALING = 0.75f; 
 	public static final float DEADZONE = 0.1f;
 }
